+++ example/myshop/admin/__init__.py
@@ -2,11 +2,9 @@
 from __future__ import unicode_literals
 
 from django.conf import settings
-from django.contrib import admin
-from shop.admin.customer import CustomerProxy, CustomerAdmin
-from shop.models.order import OrderModel
 
-__all__ = ['OrderModel', 'commodity']
+from shop.admin.defaults import customer
+from shop.models.order import OrderModel
 
 # models defined by the myshop instance itself
 if settings.SHOP_TUTORIAL in ('commodity', 'i18n_commodity'):
@@ -24,4 +22,4 @@ elif settings.SHOP_TUTORIAL == 'polymorphic':
     from . import manufacturer
     from .polymorphic import product, order
 
-admin.site.register(CustomerProxy, CustomerAdmin)
+__all__ = ['OrderModel', 'commodity', 'customer']